bad convertion solved

+++ src_files/src/src_server/server_game/server_unidades/devastador.cpp
@@ -9,7 +9,7 @@ Devastador::Devastador(uint8_t id, Jugador& duenio, Mapa& mapa, YAML::Node& atri
     this->velocidad = atributos_unidad["Vehiculo"]["Devastador"]["Velocidad"].as<int16_t>();
     this->tiempo_entrenamiento = atributos_unidad["Vehiculo"]["Devastador"]["TiempoEntrenamiento"].as<uint16_t>();
     this->vida = atributos_unidad["Vehiculo"]["Devastador"]["Vida"].as<int16_t>();
-    this->costo = atributos_unidad["Vehiculo"]["Devastador"]["Costo"].as<uint8_t>();
+    this->costo = atributos_unidad["Vehiculo"]["Devastador"]["Costo"].as<uint16_t>();
     
     std::vector<char> aux = atributos_unidad["Vehiculo"]["Devastador"]["PenalizacionTerreno"].as<std::vector<char>>();
     std::vector<float> aux_float = atributos_unidad["Vehiculo"]["Devastador"]["PenalizacionVelocidad"].as<std::vector<float>>();
