+++ src_files/src/src_server/server_game/server_unidades/fremen.cpp
@@ -9,7 +9,7 @@ Fremen::Fremen(uint8_t id, Jugador& duenio, Mapa& mapa, YAML::Node& atributos_un
     this->velocidad = atributos_unidad["Infanteria"]["Fremen"]["Velocidad"].as<int16_t>();
     this->tiempo_entrenamiento = atributos_unidad["Infanteria"]["Fremen"]["TiempoEntrenamiento"].as<uint16_t>();
     this->vida = atributos_unidad["Infanteria"]["Fremen"]["Vida"].as<int16_t>();
-    this->costo = atributos_unidad["Infanteria"]["Fremen"]["Costo"].as<uint8_t>();
+    this->costo = atributos_unidad["Infanteria"]["Fremen"]["Costo"].as<uint16_t>();
     
     std::vector<char> aux = atributos_unidad["Infanteria"]["Fremen"]["PenalizacionTerreno"].as<std::vector<char>>();
     std::vector<float> aux_float = atributos_unidad["Infanteria"]["Fremen"]["PenalizacionVelocidad"].as<std::vector<float>>();
