+++ src_files/src/src_server/server_game/server_unidades/infanteriaLigera.cpp
@@ -13,7 +13,7 @@ InfanteriaLigera::InfanteriaLigera(uint8_t id, Jugador& duenio, Mapa& mapa, YAML
     this->velocidad = atributos_unidad["Infanteria"]["InfanteriaLigera"]["Velocidad"].as<int16_t>();
     this->tiempo_entrenamiento = atributos_unidad["Infanteria"]["InfanteriaLigera"]["TiempoEntrenamiento"].as<uint16_t>();
     this->vida = atributos_unidad["Infanteria"]["InfanteriaLigera"]["Vida"].as<int16_t>();
-    this->costo = atributos_unidad["Infanteria"]["InfanteriaLigera"]["Costo"].as<uint8_t>();
+    this->costo = atributos_unidad["Infanteria"]["InfanteriaLigera"]["Costo"].as<uint16_t>();
 
     std::vector<char> aux = atributos_unidad["Infanteria"]["InfanteriaLigera"]["PenalizacionTerreno"].as<std::vector<char>>();
     std::vector<float> aux_float = atributos_unidad["Infanteria"]["InfanteriaLigera"]["PenalizacionVelocidad"].as<std::vector<float>>();
