+++ src_files/src/src_server/server_game/server_unidades/tanqueSonico.cpp
@@ -9,7 +9,7 @@ TanqueSonico::TanqueSonico(uint8_t id, Jugador& duenio, Mapa& mapa, YAML::Node& 
     this->velocidad = atributos_unidad["Vehiculo"]["TanqueSonico"]["Velocidad"].as<int16_t>();
     this->tiempo_entrenamiento = atributos_unidad["Vehiculo"]["TanqueSonico"]["TiempoEntrenamiento"].as<uint16_t>();
     this->vida = atributos_unidad["Vehiculo"]["TanqueSonico"]["Vida"].as<int16_t>();
-    this->costo = atributos_unidad["Vehiculo"]["TanqueSonico"]["Costo"].as<uint8_t>();
+    this->costo = atributos_unidad["Vehiculo"]["TanqueSonico"]["Costo"].as<uint16_t>();
     
     std::vector<char> aux = atributos_unidad["Vehiculo"]["TanqueSonico"]["PenalizacionTerreno"].as<std::vector<char>>();
     std::vector<float> aux_float = atributos_unidad["Vehiculo"]["TanqueSonico"]["PenalizacionVelocidad"].as<std::vector<float>>();
